Fix trying to open the sidebar when not available
When OCA.Files is not defined the sidebar is not available, but in that
case it is not possible to check against OCA.Files.Sidebar.

When the sidebar is not available "sidebarFile" is null, and when the
sidebar is available but there is no file it is an empty string, so it
is enough to check if "sidebarFile" has a value or not in order to open
the sidebar or not.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>
Signed-off-by: npmbuildbot[bot] <npmbuildbot[bot]@users.noreply.github.com>