Fix fullscreen and button style
":deep()" was introduced in Vue 3, although it can be used in Vue 2 too
if vue-loader >= 5.10.0. However, in stable25 the viewer is using Vue 2
with vue-loader 5.9.8, so ":deep()" has no effect and therefore the
styles for fullscreen mode and buttons were not applied.

Given that the change from "::v-deep" to ":deep()" was done in a recent
backport and that, except for security fixes, the dependencies are not
bumped in stable versions "::v-deep" is used again instead of bumping
the vue-loader version.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>
Signed-off-by: nextcloud-command <nextcloud-command@users.noreply.github.com>