fix: Load preview URL for download-disabled shares
This was possible on Nextcloud 30 and previous due to a "bug":
The `download` permission was simply not rejected for public shares,
just a `hide` flag was set on the public share.

Now the permissions are correctly set, so loading a preview is not possible.
The work-around is to allow previews when the correct header is set.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>